Update and fix linters errors
