+++ src/logic/task.js
@@ -1,0 +1,11 @@
+// Task class : represents one task
+
+class Task {
+  constructor(description, completed, index) {
+    this.description = description;
+    this.completed = completed;
+    this.index = index;
+  }
+}
+
+export default Task;
